requirements.txt
numpy
torch
transformers
tokenizers
accelerate
trl
peft
sentencepiece
bitsandbytes
scipy
datasets
pyyaml
torch
